% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_metrics.R
\name{bin_scored_metrics}
\alias{bin_scored_metrics}
\title{Binning scored climate metrics}
\usage{
bin_scored_metrics(metric_df, score_result, bins)
}
\arguments{
\item{metric_df}{A data frame of metrics calculated for each model run. For example,
an output from \code{\link{metric_calc}}.}

\item{score_result}{A data frame of scores calculated for each model run. For example,
an output from \code{\link{score_hruns}}.}

\item{bins}{A vector of variable ranges used to bin metric data.}
}
\value{
A data frame object of bins and the summed scores of model runs producing
metrics in the value range of each bin.
}
\description{
This function organizes metric data into bins and computes the sum of the
scored runs in each bin. This method effectively weights bins by how closely each run
represents observed data.
}
