% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_pdf_sample.R
\name{covariance_matrix}
\alias{covariance_matrix}
\title{Building Variance-Covariance Matrix}
\usage{
covariance_matrix(dist1, dist2, param_cor = -0.75)
}
\arguments{
\item{dist1}{Probability distribution of the first covariate.}

\item{dist2}{Probability distribution of the second covariate.}

\item{param_cor}{Correlation of the covariates contributing to the joint
probability distribution.}
}
\value{
A matrix object with diagonal elements containing the variances of the
variables and off-diagonal elements containing variance of the paired variables.
}
\description{
This function builds a variance-covariance matrix that contains
the variances and covariances of associated with provided variable distributions.
}
