% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_crit.R
\name{crit_CO2_obs}
\alias{crit_CO2_obs}
\title{Screening criterion using Mauna Loa atmospheric CO2}
\usage{
crit_CO2_obs()
}
\value{
A criterion using Mauna Loa atmospheric CO2
}
\description{
Screening criterion using Mauna Loa atmospheric CO2
}
\examples{
crit_CO2_obs()
}
