% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_criterion.R
\name{criterion_co2_obs}
\alias{criterion_co2_obs}
\title{Screening criterion using Mauna Loa atmospheric CO2}
\usage{
criterion_co2_obs()
}
\value{
A criterion identifier using Mauna Loa atmospheric CO2
}
\description{
This is a criterion identifier for screening Hector runs using
observed CO2 levels (ppm) from the Mauna Loa data set. The data set provides the
longest running measurements of mean annual atmospheric CO2, collected from the
Mauna Loa Observatory.
}
\note{
This function uses the Mauna Loa data set from
\href{https://gml.noaa.gov/ccgg/trends/data.html}{https://gml.noaa.gov/ccgg/trends/data.html}
downloaded on 01/03/2023
}
\examples{
criterion_co2_obs()
}
