% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_criterion.R
\name{criterion_historic_co2}
\alias{criterion_historic_co2}
\title{Screening criterion using historic CO2 concentrations}
\usage{
criterion_historic_co2()
}
\value{
A criterion identifier using historic co2 concentrations
}
\description{
This is a criterion identifier for weighting a Matilda ensemble using
historic CO2 concentration. The data used for this criterion are from the Indicators of
Global Climate Change paper (2023).
}
\note{
For full description, citation, and references see \link{historic_co2_concentrations}
}
\examples{
criterion_historic_co2()
}
\seealso{
\link{historic_co2_concentrations}
}
