% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_criterion.R
\name{criterion_mlo_co2}
\alias{criterion_mlo_co2}
\title{Screening criterion using CO2 concentrations from Mauna Loa Observatory (MLO)}
\usage{
criterion_mlo_co2()
}
\value{
A criterion identifier using MLO atmospheric CO2 concentration data.
}
\description{
This is a criterion identifier for screening Matilda ensembles using
observed atmospheric CO2 concentrations (ppm) from the Mauna Loa Observatory (MLO).
The data set provides the longest running observations of mean annual atmospheric CO2 concentrations.
}
\note{
This function uses the Mauna Loa data set from
\href{https://gml.noaa.gov/ccgg/trends/data.html}{https://gml.noaa.gov/ccgg/trends/data.html}
downloaded on 01/03/2023
}
\examples{
criterion_mlo_co2()
}
