% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_criterion.R
\name{criterion_ocean_c_uptake_obs}
\alias{criterion_ocean_c_uptake_obs}
\title{Screening criterion using Global Carbon Project ocean carbon uptake}
\usage{
criterion_ocean_c_uptake_obs()
}
\value{
A criterion identifier using GCP ocean carbon uptake
}
\description{
This is a criterion identifier for screening Hector runs using
observed ocean carbon uptake from the Global Carbon Project's 2023 Global Carbon Budget.
}
\note{
This function uses ocean carbon uptake data from
\href{ https://globalcarbonbudgetdata.org/latest-data.html}{ https://globalcarbonbudgetdata.org/latest-data.html}
}
\examples{
criterion_ocean_c_uptake_obs()
}
