% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_criterion.R
\name{criterion_tas_obs}
\alias{criterion_tas_obs}
\title{Screening criterion using HadCRUT5 mean global temperature anomaly}
\usage{
criterion_tas_obs()
}
\value{
A criterion identifier using HadCRUT5 global temperature anomaly
}
\description{
This is a criterion identifier for screening Hector runs using
observed global temperature anomaly (C) from the HadCRUT5 data set. These data
are presented as temperature anomalies relative to 1961-1990.
}
\note{
This function uses the HadCRUT global annual mean temperature anomaly data from
\href{https://www.metoffice.gov.uk/hadobs/hadcrut5/data/current/download.html}{https://www.metoffice.gov.uk/hadobs/hadcrut5/data/current/download.html}
}
\examples{
criterion_tas_obs()
}
