% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_param.R
\name{gen_param}
\alias{gen_param}
\title{Generating Parameter Values}
\usage{
gen_param(draws)
}
\arguments{
\item{draws}{Number of random draws for each parameter.}
}
\value{
A data frame object with a row for each parameter value generated for each
run. Columns indicate run_number and each parameter.
}
\description{
Generates parameter values from a normal distribution user
provided number of random draws.
}
