% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector_matrix.R
\name{hector_matrix}
\alias{hector_matrix}
\title{Converting Hector Result to Matrix}
\usage{
hector_matrix(df, columns = "value")
}
\arguments{
\item{df}{Result data frame from \code{\link{iterate_hector}}.}

\item{columns}{Column name to select for inclusion in the matrix. Defaults to
"value" which will use the values for each model iteration.}
}
\value{
A matrix with columns representing a vector of values for each model
iteration.
}
\description{
Converting Hector Result to Matrix
}
\examples{
# Matrix hector result using model values
hector_matrix(hector_result, columns = "value")
}
