% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_crit.R
\name{is.crit}
\alias{is.crit}
\title{Is an object of type criterion?}
\usage{
is.crit(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
TRUE if x has an attribute class of criterion, and FALSE otherwise.
}
\description{
Checks whether the supplied argument has an attribute class of criterion
}
\examples{
x <- crit_co2_obs()
is.crit(x)
}
