% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_criterion.R
\name{is.criterion}
\alias{is.criterion}
\title{Is an object of type criterion?}
\usage{
is.criterion(x)
}
\arguments{
\item{x}{an \code{R} object.}
}
\value{
\code{is.crit} returns \code{TRUE} or \code{FALSE} depending on whether \code{x}
has an attribute class of \code{"criterion"} or not.
}
\description{
Checks whether \code{x} has an attribute class of \code{"criterion"}
}
\examples{
x <- criterion_co2_obs()

# Is x of class 'criterion'?
is.criterion(x)
}
