% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_hector.R
\name{iterate_hector}
\alias{iterate_hector}
\title{Iterate Hector Runs}
\usage{
iterate_hector(core, metric, params)
}
\arguments{
\item{core}{A core object to initiate Hector runs.}

\item{metric}{An object identifying a variable, year range, and operation
(e.g. mean, median, max, min, etc.) to fetch from Hector result.}

\item{params}{A data frame object containing parameter values.}
}
\value{
A data frame with a run_number from one to the total number of Hector
runs completed and values for the variables and year range identified in
the metric argument for each Hector run.
}
\description{
Runs Hector in an iterative process with parameter uncertainty.
}
\examples{
# Load scenario file and initiate a new Hector core
ssp245 <- system.file("input/hector_ssp245.ini", package = "hector")
core <- newcore(ssp245)

# Create and new metric
metric <- new_metric(GLOBAL_TAS(), years = 2000:2100, op = mean)
print(metric)

# Compute parameter values for Hector iterations
params <- generate_params(10)
params

# Iterate Hector runs with parameter uncertainty
h_result <- iterate_hector(core, metric, params)
head(h_result)
}
