% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_hector.R
\name{iterative_hector}
\alias{iterative_hector}
\title{Iterative Hector Runs}
\usage{
iterative_hector(core, var, years, params)
}
\arguments{
\item{core}{A core object to initiate Hector runs.}

\item{var}{A variable name.}

\item{years}{A year range.}

\item{runs}{A numeric value indicating the number of Hector runs to complete.}
}
\value{
A data frame with added columns indicating parameter values used for
each Hector run, metric value for each Hector run, and a run_number from one
to the total number of Hector runs completed.
}
\description{
Runs Hector in an iterative process with parameter uncertainty.
}
