% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it_hector_with_subset.R
\name{lognorm}
\alias{lognorm}
\title{Log norm calc}
\usage{
lognorm(m, sd)
}
\arguments{
\item{m}{mean}

\item{sd}{standard deviation}
}
\value{
mean and standard deviation parameters that can be used for appropriate calculation of a log normal distribution for random draws.
}
\description{
Log norm calc
}
