% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_pdf_sample.R
\name{lognorm_to_norm}
\alias{lognorm_to_norm}
\title{Converting lognormal distribution to normal distribution}
\usage{
lognorm_to_norm(n, mean, sd)
}
\arguments{
\item{n}{Number of observations.}

\item{mean}{mean of the distribution}

\item{sd}{deviation of the distribution}
}
\value{
A vector list of randomly sample values. The length of the results is
determined by \code{n}. Lognormal distribution in this function is in reference
to the result produced by \code{\link{lognorm}}.
}
\description{
Converting lognormal distribution to normal distribution
}
