% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hush.R
\name{matilda_message}
\alias{matilda_message}
\title{Suppressing messages in Matilda functions}
\usage{
matilda_message(...)
}
\arguments{
\item{...}{Message to print if verbosity option is set to TRUE.}
}
\description{
Suppressing messages in Matilda functions
}
