% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hush.R
\name{matilda_warning}
\alias{matilda_warning}
\title{Suppressing warnings in Matilda functions}
\usage{
matilda_warning(...)
}
\arguments{
\item{...}{Warning to print if verbosity option is set to TRUE.}
}
\description{
Suppressing warnings in Matilda functions
}
