% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_criteria_weighting.R
\name{multi_criteria_weighting}
\alias{multi_criteria_weighting}
\title{Weighting ensemble members using multiple criterion}
\usage{
multi_criteria_weighting(scores_list, criterion_weights = NULL)
}
\arguments{
\item{scores_list}{A list of data frames containing weights (scores) computed using
different scoring criteria.}

\item{criterion_weights}{A vector list of weights to be applied to data frame of scores
based on importance of each criterion. These weights default to apply uniform weight
across scoring criteria.}
}
\description{
Weighting ensemble members using multiple criterion
}
