% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_criterion.R
\name{new_criterion}
\alias{new_criterion}
\title{Creating New Criterion Object}
\usage{
new_criterion(var, years, obs_values)
}
\arguments{
\item{var}{A variable from Hector output to compare with observed data}

\item{years}{A range of years from observed data}

\item{obs_values}{Numeric values from observed data}
}
\value{
Pretty print of R object containing new criterion parameters.
}
\description{
This function will create a new criterion used to screen Hector
runs for realism. User can load data for comparing Hector variable outputs to
observed data values.
}
\examples{
# Assign observed data to an object
my_criterion <- new_criterion(GLOBAL_TAS(),
  years = 1951:2000,
  obs_values = seq(0.4, 1.0, length.out = 50)
)

# View printed criterion
my_criterion
}
