% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_crit.R
\name{print.crit}
\alias{print.crit}
\title{Print new criterion}
\usage{
\method{print}{crit}(x, ...)
}
\arguments{
\item{x}{Criterion object created in \code{\link{new_crit}}}

\item{...}{}
}
\value{
Pretty print of criterion stored in R object.
}
\description{
Print new criterion
}
