% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_crit.R
\name{print.criterion}
\alias{print.criterion}
\title{Pretty print of new criterion}
\usage{
\method{print}{criterion}(x, ...)
}
\arguments{
\item{x}{Criterion object created in \code{\link{new_crit}}}

\item{...}{}
}
\value{
Pretty print of criterion stored in R object.
}
\description{
Pretty print of new criterion
}
