% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_metric.R
\name{print.h_metric}
\alias{print.h_metric}
\title{Printing new metric}
\usage{
\method{print}{h_metric}(x, ...)
}
\arguments{
\item{x}{R object storing class attributes.}

\item{...}{Other arguments passed to or from other methods.}
}
\value{
Pretty print of metrics stored in R object.
}
\description{
Prints Hector metric objects produced using.
\code{\link{new_metric}}
}
