% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_calc}
\alias{prob_calc}
\title{Calculating probabilities of climate metric outcomes}
\usage{
prob_calc(x)
}
\arguments{
\item{x}{A data frame of scored frequencies of each model run organized into data bins.
For example, an output from \code{\link{bin_scored_metrics}}.}
}
\value{
A data frame of probabilities using breaks assigned by user in
\code{\link{bin_scored_metrics}}
}
\description{
This is a function that calculates basic probabilities of metric outcomes
from multiple model runs.
}
