% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE_calc.R
\name{RMSE_calc}
\alias{RMSE_calc}
\title{Calculating Root Mean Square Error (RMSE)}
\usage{
RMSE_calc(x, y, sigma = sd(y))
}
\arguments{
\item{x}{A vector of modeled data values}

\item{y}{A vector of observed data values}

\item{sigma}{A single value or vector of error terms the same length as y.
A single value will apply a constant error term. User can provide a vector
of error terms to incorporate time-varying error into the RMSE calculation.
`sigma` default assume homoscedasticity of residuals and applies a constant
error term equal to the standard deviation of observed data (scoring criterion).}
}
\value{
Returns a vector of RMSE values
}
\description{
Function takes vectors of modeled data values and compares them
to observed data values to report the RMSE.
}
\examples{
x <- c(1:5)
y <- c(5:9)

RMSE_calc(x, y)
}
