% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE_values.R
\name{rmse_values}
\alias{rmse_values}
\title{Computing Root Mean Square Error (RMSE) Values for Ensemble Members}
\usage{
rmse_values(df, criterion, sigma = NULL)
}
\arguments{
\item{df}{Result data frame from \code{\link{iterate_model}}.}

\item{criterion}{A scoring criterion to use for screening Hector runs (e.g., \code{\link{criterion_gmst_obs}})}

\item{sigma}{A single value or vector of error terms the same length as y.
A single value will apply a constant error term. User can provide a vector
of error terms to incorporate time-varying error into the RMSE calculation.
`sigma` default assume homoscedasticity of residuals and applies a constant
error term equal to the standard deviation of observed data (scoring criterion).}
}
\value{
Returns a data frame with a \code{run_number} each ensemble member and
a corresponding RMSE value.
}
\description{
This function computes RMSE values for each perturbed parameter ensemble member. It uses the result
of \code{\link{iterate_model}} to compute RMSE using the user specified criterion. Time varying error can be applied
using \code{sigma}, which supplies a single value or a vector of values representing criterion error at each time step.
}
