% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_CO2.R
\name{score_CO2}
\alias{score_CO2}
\title{Screen Hector outputs by observed CO2}
\usage{
score_CO2(x, w1 = 2, w2 = 20)
}
\arguments{
\item{x}{result data frame from iterative_hector}

\item{w1}{low divergence bound to assign highest score}

\item{w2}{high divergence bound to assign lowest score}
}
\value{
Data frame with mean score for each Hector run
}
\description{
Screen Hector outputs by observed CO2
}
