% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{set_params}
\alias{set_params}
\title{Set Parameter Values to Hector Variables}
\usage{
set_params(core, param_values)
}
\arguments{
\item{core}{Hector core object}

\item{param_values}{An object of generated parameter values. Must be of
numeric class. Column names must match parameter functions in Hector.}
}
\value{
Message indicating the values set to each parameter variable.
}
\description{
Takes randomly generated parameter values produced by
\code{\link{generate_params}} and sets each value to appropriate Hector variable
for the supplied Hector core.
}
