% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_metrics.R
\name{bin_metrics}
\alias{bin_metrics}
\title{Binning un-scored climate metrics}
\usage{
bin_metrics(metric_df, bins)
}
\arguments{
\item{metric_df}{Data frame output from \code{\link{metric_calc}} result.}

\item{bins}{a vector of breaks for data binning.}
}
\value{
A table object that bins metric values using breaks supplied
by user.
}
\description{
A function that accepts data frame of climatic metrics from
\code{\link{metric_calc}} output and groups metric values in user defined
bins.
}
