% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_metrics.R
\name{bin_scored_metrics}
\alias{bin_scored_metrics}
\title{Binning scored climate metrics and calculating probabilities}
\usage{
bin_scored_metrics(metric_df, bins, score_result)
}
\arguments{
\item{metric_df}{A data frame of metrics calculated for each model run. For example,
an output from \code{\link{metric_calc}}.}

\item{bins}{A vector of variable ranges used to bin metric data.}

\item{score_result}{A data frame of scores calculated for each model run. For example,
an output from \code{\link{score_hruns}}.}
}
\value{
A data frame object of bins and the summed scores of model runs producing
metrics in the value range of each bin.
}
\description{
This function organizes metric data into bins and computes the sum of the
scored runs in each bin. This method effectively weights bins by how closely each run
represents observed data, and calculates the probability that a metric will occur in
each bin range.
}
\examples{
# Example of data frame containing metric values
metric_df <- data.frame(run_number = 1:10,
                        metric_result = runif(10, 690, 805))
metric_df

# Example of data frame containing scores for each run
scores_df <- data.frame(run_number = 1:10,
                        scores = runif(10, 0, 1))
scores_df

# Creating bins using variable ranges
bins <- c(690, 720, 740, 760, 780, 805)

# Calculating probabilities for each bin
bin_scored_metrics(metric_df = metric_df, bins = bins, score_result = scores_df)
}
