% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_crit.R
\name{crit_co2_obs}
\alias{crit_co2_obs}
\title{Screening criterion using Mauna Loa atmospheric CO2}
\usage{
crit_co2_obs()
}
\value{
A criterion using Mauna Loa atmospheric CO2
}
\description{
This data set gives the longest running measurements of mean annual
atmospheric CO2 levels (ppm) from the Mauna Loa Observatory.
}
\note{
This uses the Mauna Loa data set from \link{https://gml.noaa.gov/ccgg/trends/data.html}
downloaded on 01/03/2023
}
\examples{
crit_co2_obs()
}
