% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_crit.R
\name{crit_co2_obs}
\alias{crit_co2_obs}
\title{Screening criterion using Mauna Loa atmospheric CO2}
\usage{
crit_co2_obs()
}
\value{
A criterion using Mauna Loa atmospheric CO2
}
\description{
Screening criterion using Mauna Loa atmospheric CO2
}
\note{
This uses the Mauna Loa on 'URL' on 'DATE'
}
\examples{
crit_co2_obs()
}
