% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_hector_result.R
\docType{data}
\name{hector_result}
\alias{hector_result}
\title{Example Hector result data}
\format{
hector_result is a data frame with 556 rows and 5 columns:
\describe{
  \item{scenario}{Scenario being used for the analysis}
  \item{year}{Years}
  \item{variable}{Variables}
  \item{value}{Value of variable}
  \item{units}{Units of the variable}
}
}
\usage{
hector_result
}
\description{
This is a data set representing the results of an output from
a single hector. The model results include the GMST() Hector variable for
the years 1745:2300.
}
\keyword{datasets}
