% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_crit.R
\name{is.crit}
\alias{is.crit}
\title{Is an object of type criterion?}
\usage{
is.crit(x)
}
\arguments{
\item{x}{an \code{R} object.}
}
\value{
\code{is.crit} returns \code{TRUE} or \code{FALSE} depending on whether \code{x}
has an attribute class of \code{"criterion"} or not.
}
\description{
Checks whether \code{x} has an attribute class of \code{"criterion"}
}
\examples{
x <- crit_co2_obs()

# Is x of class 'criterion'?
is.crit(x)
}
