% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_pdf_sample.R
\name{joint_pdf_sample}
\alias{joint_pdf_sample}
\title{Build and Sample Joint Probability Distribution}
\usage{
joint_pdf_sample(draws, dist1, dist2, param_cor = -0.75)
}
\arguments{
\item{draws}{Number of random draws for each parameter.}

\item{dist1}{Probability distribution of the first covariate.}

\item{dist2}{Probability distribution of the second covariate.}

\item{param_cor}{Correlation of the covariates contributing to the joint
probability distribution.}
}
\value{
A data frame object with parameter values generated for each
draw. Values randomly sample from a multivariate normal distribution of
correlated Hector parameters.
}
\description{
A joint probability distribution is implemented for parameters that
are correlated. A default correlation of -0.75 is used, but is specific to the
inversely correlated parameters of \code{\link{ECS}} and \code{\link{DIFFUSIVITY}}.
This function builds a variance-covariance matrix to randomly sample from a multivariate
distribution \href{link}{(Pressburger et al. 2023)}.
}
\details{
This function is currently specific to the joint PDF of \code{\link{ECS}} and
\code{\link{DIFFUSIVTIY}} and therefore uses a default parameter correlation
specific to these parameters. The user can also define their own parameter
correlation \code{param_cor}. Correlation value is multiplied by the standard
deviation of each parameter contributing to the joint probability distribution.
The function \code{\link{mvnorm}} in the package \code{\link{MASS}} was used
to randomly sample the variance-covariance matrix.
}
