% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_hector.R
\name{lognorm}
\alias{lognorm}
\title{Log normal parameterization}
\usage{
lognorm(m, sd)
}
\arguments{
\item{m}{mean}

\item{sd}{standard deviation}
}
\value{
Mean and standard deviation parameters that can be used for appropriate
calculation of a log normal distribution for random draws.
}
\description{
Function used to supply proper values to rlnorm() in order to produce
a log-normal distribution when arithmetic mean and standard deviation of a parameter are
supplied.
}
