% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_calc.R
\name{metric_calc}
\alias{metric_calc}
\title{Metric Calculation}
\usage{
metric_calc(x, metric)
}
\arguments{
\item{x}{A data frame object from \code{\link{iterate_model}} output.}

\item{metric}{An object identifying a variable, year range, and operation
(e.g. mean, median, max, min, etc.) to fetch from Hector result.}
}
\value{
A data frame with columns for \code{run_number} and
 \code{metric_result} values for each Hector run.
}
\description{
This is a function that calculates metric values for multiple
Hector runs. Metrics to be calculated from Hector runs are defined using
\code{\link{new_metric}}.
}
\examples{
# Create new metric
metric <- new_metric(var = "global_tas", years = 2000:2100, op = mean)

# Calculate metric values for each Hector run
metric_calc(matilda_result, metric)
}
