% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_calc-function-dev.R
\name{metric_calc}
\alias{metric_calc}
\title{Metric Calculation}
\usage{
metric_calc(x, op, var, years)
}
\arguments{
\item{x}{A data frame object from iterative_hector() output.}

\item{op}{An operation to be applied to the climatic variable values (e.g.,
mean, median, max, min).}

\item{var}{A character string indicating which variable to apply the operation.}

\item{years}{A year range.}
}
\value{
A data frame with columns for run_number, variable, and a metric
calculation for each Hector run.
}
\description{
A function that accepts a data frame result from an iterative Hector
run and calculates a specified metric for each run.
}
