% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_calc.R
\name{metric_calc}
\alias{metric_calc}
\title{Metric Calculation}
\usage{
metric_calc(x, metric)
}
\arguments{
\item{x}{A data frame object from \code{\link{iterate_hector}} output.}

\item{metric}{An object identifying a variable, year range, and operation
(e.g. mean, median, max, min, etc.) to fetch from Hector result.}
}
\value{
A data frame with columns for run_number, variable, and a metric
calculation for each Hector run.
}
\description{
A function that accepts a data frame result from an iterative Hector
run and calculates a specified metric for each run.
}
\examples{
# Load scenario file and initiate a new hector core
ssp245 <- system.file("input/hector_ssp245.ini", package = "hector")
core <- newcore(ssp245)

# Create new metric
metric <- new_metric(GLOBAL_TAS(), years = 2000:2100, op = mean)

# Compute parameter values for Hector iterations
params <- generate_params(10)

# Iterate Hector runs with parameter uncertainty
h_result <- iterate_hector(core, metric, params)

# Calculate metric values for each Hector run
metric_calc(h_result, metric)
}
