% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_hector.R
\name{metric_calc_1run}
\alias{metric_calc_1run}
\title{Metric calc from single Hector Run}
\usage{
metric_calc_1run(x, metric)
}
\arguments{
\item{x}{A data frame result from a single Hector run.}

\item{metric}{An object identifying a variable, year range, and operation
(e.g. mean, median, max, min, etc.) to apply to data.}
}
\value{
A numeric value calculated from the operation for each variable in the
year range.
}
\description{
Function for calculating a variable metric from Hector output data.
}
