% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_hector.R
\name{metric_calc_1run}
\alias{metric_calc_1run}
\title{Metric calc from single Hector Run}
\usage{
metric_calc_1run(df_1run, op, var, years)
}
\arguments{
\item{df_1run}{A data frame result from a single Hector run.}

\item{op}{An operation to apply to data (e.g. mean, median, max, min, etc.).}

\item{var}{A variable name.}

\item{years}{A year range.}
}
\value{
A numeric value calculated from the operation for each variable in the
year range.
}
\description{
Function for calculating a variable metric from Hector output data.
}
