% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_criteria_weighting.R
\name{multi_criteria_weighting}
\alias{multi_criteria_weighting}
\title{Weighting ensemble members using multiple criterion}
\usage{
multi_criteria_weighting(scores_list, criterion_weights = NULL)
}
\arguments{
\item{scores_list}{A list of data frames containing weights (scores) computed using
different scoring criteria.}

\item{criterion_weights}{A vector list of weights to be applied to data frame of scores
based on importance of each criterion. These weights default to apply uniform weight
across scoring criteria.}
}
\value{
A data frame with a column of run_numbers and a column of weights computed
using multiple scoring criterion (mc_weight).
}
\description{
Weighting ensemble members using multiple criterion
}
\examples{
# Data frames representing scored ensemble using different scoring criterion
score_df1 <- data.frame(run_number = c(1, 2, 3), weights = c(0.8, 0.6, 0.9))
score_df2 <- data.frame(run_number = c(1, 2, 3), weights = c(0.7, 0.5, 0.8))
score_df3 <- data.frame(run_number = c(1, 2, 3), weights = c(0.9, 0.8, 0.7))

# List of score data frames
scores_list <- list(score_df1, score_df2, score_df3)

# Define weights for each criterion
criterion_weights <- c(0.4, 0.3, 0.3)

# Call the multi_criteria_weighting function
result <- multi_criteria_weighting(scores_list, criterion_weights)

# Resulting data frame
result

}
