% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_metric.R
\name{new_metric}
\alias{new_metric}
\title{Creating new metric object}
\usage{
new_metric(var, years, op)
}
\arguments{
\item{var}{A character string of Hector variable}

\item{years}{A range of years}

\item{op}{An operation to apply to data}
}
\value{
Pretty print of R object containing new metric values.
}
\description{
Stores metric values in an object for easily interchanging new
metrics for functions within hectorpractice.
}
