% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_crit.R
\name{print.crit}
\alias{print.crit}
\title{Print new criterion}
\usage{
\method{print}{crit}(x, ...)
}
\arguments{
\item{x}{Criterion object created in \code{\link{new_crit}}}

\item{...}{Other arguments to or from other methods.}
}
\value{
Printed version of criterion stored in \code{R} object.
}
\description{
This prints the screening criterion provided as \code{x}.
}
\examples{
print.crit(crit_co2_obs())
}
