% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_metric.R
\name{print.h_metric}
\alias{print.h_metric}
\title{Printing new metric}
\usage{
\method{print}{h_metric}(x, ...)
}
\arguments{
\item{x}{An \code{R} object storing the Hector metric.
Built using \code{\link{new_metric}}.}

\item{...}{Other arguments passed to or from other methods.}
}
\value{
Printed version of metrics stored in \code{R} object.
}
\description{
Prints the Hector metric provided as \code{x}.
}
