% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_calc}
\alias{prob_calc}
\title{Calculating probabilities of climate metric outcomes}
\usage{
prob_calc(x)
}
\arguments{
\item{x}{Table of binned climate variable metrics.}
}
\value{
Table of probabilities using breaks assigned by user in
\code{\link{bin_metrics}}
}
\description{
A function that calculates basic probabilities of outcomes based
on previously defined bin groupings.
}
