% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_calc}
\alias{prob_calc}
\title{Binning scored climate metrics and calculating probabilities}
\usage{
prob_calc(metrics, bins, scores = rep(1, length(metrics)))
}
\arguments{
\item{metrics}{A vector of metrics calculated for each model run. For
example, an output from \code{\link{metric_calc}}.}

\item{bins}{A vector of variable ranges used to bin metric data.}

\item{scores}{A vector of scores calculated for each model run. For
example, an output from \code{\link{score_runs}}.}
}
\value{
A data frame object of bins, the summed scores of model runs producing
metrics in the value range of each bin, and the probability of each bin outcome.
}
\description{
This function organizes metric data into bins and computes the
sum of the scored Hector runs in each bin. This method effectively weights bins
by how closely each run represents observed data. This function also calculates
the probability that a metric will occur in each bin range.
}
\examples{
# Example vector containing metric values
metrics <- runif(5, 690, 805)
metrics

# Example vector containing scores for each run
scores <- runif(5, 0, 1)
scores

# Creating bins using variable ranges
bins <- c(690, 720, 740, 760, 780, 805)

# Calculating probabilities for each bin
prob_calc(metrics = metrics, bins = bins, score = scores)
}
