% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bayesian.R
\name{score_bayesian}
\alias{score_bayesian}
\title{Computing Model Scores as Posterior Probabilities using Bayesian Inference}
\usage{
score_bayesian(m, sigma = NULL, sensitivity = NULL)
}
\arguments{
\item{m}{A Matrix of values. The first column of the matrix should be
a vector of observed data for a give variable. Subsequent vectors should be
representative of modeled values for a given variable.}

\item{sigma}{Numeric value (optional). A single value or vector of error terms
the same length as y. A single value will apply a constant error term. User
can provide a vector of error terms to incorporate time-varying error
to the RMSE calculation. `sigma` default assume homoscedasticity of residuals
and applies a constant error term equal to the standard deviation of observed
data (scoring criterion).}

\item{sensitivity}{A multiplier that adjusts the sensitivity of the likelihood
values to increasing RMSE. If not provided, the function will automatically
calculate the sensitivity as one unit of standard deviation of the RMSE results.
A smaller sensitivity value will make the Bayesian analysis give more weight
to models with lower RMSE values.}
}
\value{
Returns a vector of scores with a length equal to the number of
model iterations in the input matrix. Or, if the input matrix has K columns,
the return vector will = K - 1.
}
\description{
This function uses observed data to compute scores that represent
how well modeled values reflect what is occurring in reality. To do this, the
function uses root mean square error (RMSE) values to compute the likelihood
of observing modeled values given observed values, assuming normal distribution
of errors. Likelihood values are used to compute posterior probabilities which
are used as scores for each model iteration.
}
\note{
Note: In Bayesian statistics, the choice of `sigma` can significantly
impact the results and conclusions of the analysis. Users are encouraged to
experiment with different values and understand the implications for their
specific use case.
}
\examples{
# creating sample matrix
mat <- matrix(data = 1:20, nrow = 5, ncol = 4)

# scoring with a decay rate of 2
score_bayesian(mat, sensitivity = 2)
}
