% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bayesian.R
\name{score_bayesian}
\alias{score_bayesian}
\title{Computing Model Scores as Posterior Probabilities using Bayesian Inference}
\usage{
score_bayesian(m, e = 2, na)
}
\arguments{
\item{m}{A Matrix of values. The first column of the matrix should be
a vector of observed data for a give variable. Subsequent vectors should be
representative of modeled values for a given variable.}

\item{e}{A value from 0-Inf. This value controls the decay rate. It controls
how much models are penalized for deviation from observed data. The default is
set to 2, larger values will increase rate of decay.}
}
\value{
Returns a vector of scores with a length equal to the number of
model iterations in the input matrix. Or, if the input matrix has K columns,
the return vector will = K - 1.
}
\description{
This function uses observed data to compute scores that represent
how well modeled values reflect what is occurring in reality. To do this, the
function uses root mean square error (RMSE) values to compute the likelihood
of observing modeled values given observed values, assuming normal distribution
of errors. Likelihood values are used to compute posterior probabilities which
are used as scores for each model iteration.
}
\examples{
# creating sample matrix
mat <- matrix(data = 1:15, nrow = 5, ncol = 3)

# scoring with a decay rate of 2
score_bayesian(mat, e = 2)
}
