% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hruns.R
\name{score_hruns}
\alias{score_hruns}
\title{Screen Hector outputs with observed data}
\usage{
score_hruns(x, crit, score_function, ...)
}
\arguments{
\item{x}{result data frame from iterative_hector.}

\item{obs_dat}{observed data used to score hector runs.}

\item{years}{year range of used for screening hector runs.}

\item{w1}{low divergence bound to assign highest score.}

\item{w2}{high divergence bound to assign lowest score.}
}
\value{
Data frame with mean score for each Hector run
}
\description{
Screen Hector outputs with observed data
}
