% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_runs.R
\name{score_hruns}
\alias{score_hruns}
\title{Screen Hector outputs with observed data}
\usage{
score_hruns(x, criterion, score_function, ...)
}
\arguments{
\item{x}{Result data frame from \code{\link{iterate_hector}}.}

\item{criterion}{A scoring criterion to use for screening Hector runs.}

\item{score_function}{Scoring function to use for screening Hector model runs.}

\item{...}{Additional arguments needed to run the selected scoring function.}
}
\value{
Data frame of scored for each Hector run
}
\description{
This function uses any scoring function to screen Hector runs
based on proximity of climate variable values to observed data. Internal scoring
functions are provided in \code{matilda}, but users can also supply their own.
Criterion used for scoring are also available in the package. Alternatively,
users can build their own scoring criterion with \code{\link{new_criterion}}.
}
\examples{
# Score Hector using observed CO2 data with the score_ramp method
score_hruns(matilda_result, criterion_co2_obs(), score_ramp, w1 = 2, w2 = 20)
}
