% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_CO2.R
\name{score_ramp}
\alias{score_ramp}
\title{Computing vector of score for numeric vectors}
\usage{
score_ramp(x, y, w1, w2)
}
\arguments{
\item{x}{First set of numeric vectors}

\item{y}{Second set of numeric vectors}

\item{w1}{Distance at which score ramps down from 1.}

\item{w2}{Distance at which score becomes 0}
}
\description{
Computing vector of score for numeric vectors
}
