% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hruns.R
\name{score_ramp}
\alias{score_ramp}
\title{Computing vector of scores for numeric vectors}
\usage{
score_ramp(x, y, w1, w2, na.omit = FALSE)
}
\arguments{
\item{x}{First set of numeric vectors}

\item{y}{Second set of numeric vectors}

\item{w1}{Distance at which score ramps down from 1.}

\item{w2}{Distance at which score becomes 0}

\item{na.omit}{How should score_ramp deal with rows in data frame containing NAs?
Defaults to FALSE, leaving NAs in data frame.}
}
\value{
Vector of scores
}
\description{
Computing vector of scores for numeric vectors
}
\examples{
x <- c(3, 5, 3, 4, 9)
y <- c(3, 6.5, 5, 9, 10)
score_ramp(x, y, w1 = 1, w2 = 3)
}
