% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ramp.R
\name{score_ramp}
\alias{score_ramp}
\title{Computing Model Scores Using Ramp Method}
\usage{
score_ramp(m, w1, w2, na.omit = FALSE)
}
\arguments{
\item{m}{Matrix of values. The first column of the matrix should be
a vector of observed data. Subsequent vectors are representative of modeled
values for a given variable.}

\item{w1}{Difference at which score begins to ramp down from a 1.}

\item{w2}{Difference at which score hits 0.}

\item{na.omit}{How should \code{score_ramp} deal with rows in matrix that
contain NAs? Defaults to \code{FALSE}, leaving NAs in the matrix.}
}
\value{
Returns a vector of scores with a length equal to the number of
model iterations in the input matrix. Or, if the input matrix has K columns,
the return vector will = K - 1.
}
\description{
This function uses observed data to compute scores that represent
how well modeled values reflect what is happening in reality. Scores that are
closer to 1 indicated skillful models, while scores closer to 0 indicate models
that produce values that deviate further from observed data.
}
\examples{
# creating sample matrix
mat <- matrix(data = 1:15, nrow = 5, ncol = 3)

# scoring columns where scores <= 5 will score 1 and scores >= 10 will score 0
score_ramp(mat, w1 = 5, w2 = 10)
}
