% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{set_params}
\alias{set_params}
\title{Set Parameter Values to Hector Variables}
\usage{
set_params(core, param_values)
}
\arguments{
\item{core}{Hector core object}

\item{param_values}{An object of generated parameter values. Must be of
numeric class. Column names must match parameter functions in Hector.}
}
\value{
Message indicating the values set to each parameter variable.
}
\description{
This function takes randomly generated parameter values produced by
\code{\link{generate_params}} and sets each value Hector variables for the
supplied Hector core.
}
\examples{
ssp245 <- system.file("input/hector_ssp245.ini", package = "hector")
core <- newcore(ssp245)
params <- generate_params(core, 10)
param_vals <- unlist(params[1, ])
set_params(core, param_vals)
}
